requirements.txt
aiogram>=2.11.2,<3.0
environs~=8.0.0
xlwt>=1.3.0
amocrm_api>=2.3.1
pip~=21.0.1
PyJWT~=2.0.1
requests~=2.25.1